Reorganized structure of auditor.outbound
