+++ auditor.outbound/src/main/java/io/github/tkaczenko/auditor/outbound/api/RestTemplateAuditOutboundRequestInterceptorWithAspect.java
@@ -1,9 +1,9 @@
-package io.github.tkaczenko.auditor.outbound.resttemplate.aspect;
+package io.github.tkaczenko.auditor.outbound.api;
 
-import io.github.tkaczenko.auditor.core.internal.service.AuditDateTimeProvider;
 import io.github.tkaczenko.auditor.core.internal.AuditFacade;
 import io.github.tkaczenko.auditor.core.internal.factory.body.BodyHttpReaderService;
 import io.github.tkaczenko.auditor.core.internal.factory.headers.HeadersHttpReaderService;
+import io.github.tkaczenko.auditor.core.internal.service.AuditDateTimeProvider;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.Optional;
