+++ auditor.outbound/src/main/java/io/github/tkaczenko/auditor/outbound/api/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains the aspect used in the outbound service. The aspect is responsible for
+ * logging the request and response data for each REST call. This package contains the interceptor
+ * for the RestTemplate used in the outbound service. The service is responsible for logging the
+ * request and response data for each REST call.
+ */
+package io.github.tkaczenko.auditor.outbound.api;
