+++ auditor.outbound/src/main/java/io/github/tkaczenko/auditor/outbound/api/reader/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Provides classes and interfaces for reading data from various sources. The main entry point is
+ * the {@link io.github.tkaczenko.auditor.core.api.reader.HttpReader} interface, which defines the
+ * contract for reading data. Implementations of this interface can be used to read data from
+ * different sources, such as requests, responses for HTTP clients. The package also includes
+ * utility classes and helper methods to facilitate the reading of data.
+ */
+package io.github.tkaczenko.auditor.outbound.api.reader;
