--- auditor.outbound/src/main/java/io/github/tkaczenko/auditor/outbound/aspect/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package contains the aspect used in the outbound service. The aspect
- * is responsible for logging the request and response data for each REST call.
- */
-package io.github.tkaczenko.auditor.outbound.aspect;
