--- auditor.outbound/src/main/java/io/github/tkaczenko/auditor/outbound/resttemplate/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Provides utility classes and interfaces for interacting with external services using the Spring
- * RestTemplate. This package contains classes that handle the configuration, execution, and error
- * handling of REST API calls.
- */
-package io.github.tkaczenko.auditor.outbound.resttemplate;
